Rework sniffing mode to be more scalable
Some IR signals require pulse lengths up to 9ms, some have pause lengths
which are even longer.

The previous UART protocol would simply emit an overflow for those and
the entire thing would enter something resembling an undefined state. We
don't want that.

However, we have to stick to 8 bit messages, because even with the (now
introduced and enabled-by-default) fast mode of 100kBaud, sending a
single 8 bit message takes 100us, which is approximately four cycles of
the 38 kHz carrier, which is about the amount of time an IR decoder IC
might take to latch onto the signal. That means, we must be able to emit
frames at that rate.

One way to tackle this would be something like an UTF-8 encoding: If we
have had a particularly long pause/pulse, we use a marker bit to
indicate that this is a 14 bit value. However, this has two issues: we
lose another bit (narrowing the default range to just 6 bits of length)
and it requires us to transmit *two* UART frames, but we can only do
that *after* the long pause/pulse, which may easily be followed by a
short pulse/pause, thus making us overrun the TX buffer.

The way this commit does it is O(t) for long pulses/pauses, but has the
advantage that we do not lose any bits. It also allows us to increase
the precision to 16us (from 32us), because we don't have to live in
constant fear of overflows. In fact, 16us also has other advantages I'll
get into below.

The TIMER1 is changed to use 8us steps. We use COMPA for overflow
counting, i.e. detecting when the counter value goes beyond what we
could encode and emitting an overflow frame. Overflow frames are,
effectively, just the current count value of 0x7f plus the pause/pulse
bit. The receiver on the other end of the line then has to accumulate
the pulse/pause length. End of symbol is represented by a zero-length
pulse after a pause. These should normally not happen (it would
require a pulse which is shorter than a single 38 kHz cycle, which is
unlikely to ever be detected) and the receving end of the UART frames
should take that into account when decoding.

Using 8us steps for the TIMER1 and 16us in the UART allows us to encode
frames by just taking the seven most significant bits of the lower half
of the timer counter, slapping the pause bit onto it and calling it a
day.

Finally, the timer values for delay and hold have been overhauled.
Looking at the data from the IRMP project [1] and the times there, a
hold time of 16ms was chosen. That makes this project incompatbile
with some of the more exotic IR codes out there, but this is okay; NEC
and DENON are the most important ones here anyway.

The delay timer has been reduced to just about two 38 kHz cycles. We
have to assume that the receiver is fast, otherwise we may easily get
into some troubles. Also the advantage here is that we actually have a
direct, non-optical link to the receiver, so it's very likely that it
gets a very clear signal from us, removing the need for further
prolonging of the cycles.

   [1]: https://www.mikrocontroller.net/articles/IRMP