Do not invert the IR input
We need a level shifter for the IR demodulator, because it runs on 5V,
while the ATtiny2313a will run on 3.3V (to easier interface with the
Pi).

A simple N-FET based level shifter will cause the signal to be inverted.
The TSOP1738 is active-low. That means that the input signal we receive
will be the uninverted, demodulated IR signal (active-low + inversion =
uninverted).